feat(film): add middlewares and refactor routes
This commit adds new middlewares to film module and refactors film.routes.ts file. The new middlewares include:

getFilmListMiddleware: it takes a film object from res.locals.film and returns the list of films.
addFilmMetaAttributes: it takes an object with values as functions and film list object from res.locals.filmList and adds meta data to the filmList.
updateGoogleSheetMiddleware: it takes a filmList from res.locals.filmList, creates a sheet, and pushes the data to it.
requestedMetaMiddleware: it is a customized middleware that uses the addFilmMetaAttributesMiddleware and adds the requested additional attributes.
In addition to the new middlewares, the film.routes.ts file has been refactored to include the new middlewares.

Also, this commit renames the getFilm method in FilmController to searchFilmList and adds a new method getFilm to get the film by imdbid.

Lastly, the omdbService has been refactored to return the full movie data for getMovieList method and a new method getMovie has been added to get a single movie by imdbId.